Fixed neural network

+++ src/balle/strategy/bezierNav/WheelSpeedExecutor.java
@@ -4,7 +4,7 @@ import balle.strategy.executor.Executor;
 
 public interface WheelSpeedExecutor extends Executor {
 
-	public void update(int desLeft, int desRight, double actLeft,
+	public void update(double desLeft, double desRight, double actLeft,
 			double actRight);
 
 }
